+++ sunpy/physics/tests/test_differential_rotation.py
@@ -3,7 +3,7 @@ import pytest
 import numpy as np
 from astropy import units as u
 from astropy.coordinates import Longitude, Latitude, Angle
-from sunpy.physics.transforms.differential_rotation import diff_rot, _sun_pos, _calc_P_B0_SD, rot_hpc
+from sunpy.physics.differential_rotation import diff_rot, _sun_pos, _calc_P_B0_SD, rot_hpc
 from sunpy.tests.helpers import assert_quantity_allclose
 #pylint: disable=C0103,R0904,W0201,W0212,W0232,E1103
